Moving everything to Helper and adding loading bar
